refactor: improve Transaction entity with more fields and BigDecimal precision

+++ user-service/src/main/java/com/banking/core/account/dto/AccountResponse.java
@@ -1,6 +1,7 @@
 
 package com.banking.core.account.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -8,5 +9,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class AccountResponse {
     private String accountNumber;
-    private Long balance;
+    private BigDecimal balance;
 }
