+++ user-service/src/main/java/com/banking/core/transferservice/dto/TransferResponse.java
@@ -1,6 +1,7 @@
 
 package com.banking.core.transferservice.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,6 +13,6 @@ public class TransferResponse {
     private String transactionId;
     private String fromAccount;
     private String toAccount;
-    private Long amount;
-    private Long remainingBalance;
+    private BigDecimal amount;
+    private BigDecimal remainingBalance;
 }
